pricing - dep to utils removed
